+++ custom_components/supernotify/model.py
@@ -4,7 +4,7 @@ import re
 from collections.abc import Sequence
 from dataclasses import dataclass, field
 from enum import StrEnum, auto
-from typing import Any, ClassVar, cast
+from typing import Any, ClassVar
 
 # This import brings in a bunch of other dependency noises, make it manual until py3.14/lazy import/HA updated
 # from homeassistant.components.mobile_app import DOMAIN as MOBILE_APP_DOMAIN
@@ -31,7 +31,6 @@ from . import (
     ATTR_MOBILE_APP_ID,
     ATTR_PERSON_ID,
     ATTR_PHONE,
-    CONF_APPLY,
     CONF_DATA,
     CONF_DELIVERY_DEFAULTS,
     CONF_DEVICE_DISCOVERY,
@@ -48,7 +47,6 @@ from . import (
     RE_DEVICE_ID,
     SELECTION_DEFAULT,
     TARGET_USE_ON_NO_ACTION_TARGETS,
-    CustomizationApplication,
     SelectionRank,
 )
 from .common import ensure_list
@@ -383,16 +381,7 @@ class TransportConfig:
 class DeliveryCustomization:
     def __init__(self, config: ConfigType | None, target_specific: bool = False) -> None:
         config = config or {}
-        self.apply: CustomizationApplication
-        if config.get(CONF_ENABLED) is not None and CONF_APPLY not in config:
-            enabled: bool = cast("bool", config.get(CONF_ENABLED))
-            if enabled:
-                self.apply = CustomizationApplication.ENABLE
-            else:
-                self.apply = CustomizationApplication.DISABLE
-        else:
-            self.apply = config.get(CONF_APPLY, CustomizationApplication.ENABLE)
-
+        self.enabled: bool | None = config.get(CONF_ENABLED, True)
         self.data: dict[str, Any] | None = config.get(CONF_DATA)
         self.target: Target | None
         if config.get(CONF_TARGET):
@@ -407,7 +396,7 @@ class DeliveryCustomization:
         return self.data.get(key) if self.data else None
 
     def as_dict(self, **_kwargs: Any) -> dict[str, Any]:
-        return {CONF_TARGET: self.target.as_dict() if self.target else None, CONF_APPLY: self.apply, CONF_DATA: self.data}
+        return {CONF_TARGET: self.target.as_dict() if self.target else None, CONF_ENABLED: self.enabled, CONF_DATA: self.data}
 
 
 class DeliveryConfig: