+++ tests/supernotify/test_scenario.py
@@ -9,7 +9,6 @@ from custom_components.supernotify import (
     ATTR_PRIORITY,
     ATTR_SCENARIOS_APPLY,
     ATTR_SCENARIOS_CONSTRAIN,
-    CONF_APPLY,
     CONF_DELIVERY,
     CONF_SELECTION,
     CONF_TRANSPORT,
@@ -372,7 +371,7 @@ async def test_scenario_override_only_preselected_delivery(hass: HomeAssistant) 
         homeassistant=hass,
         deliveries={"plain_email": {CONF_TRANSPORT: "dummy", CONF_SELECTION: "explicit"}, "text": {CONF_TRANSPORT: "dummy"}},
         transport_types=[DummyTransport],
-        scenarios={"Spammy": {CONF_DELIVERY: {"plain_email": {CONF_APPLY: "override", "data": {"priority": "low"}}}}},
+        scenarios={"Spammy": {CONF_DELIVERY: {"plain_email": {"enabled": None, "data": {"priority": "low"}}}}},
     )
     await ctx.test_initialize()